Partial input handling for new book creation.
As of now, the profile is working fine. Also, when user adds a new book, the names given under secondary owners are validated in real time with the Firebase DB. The UI works seamlessly until this. Book name duplicates are not yet validated. No bugs are discovered in handling this dynamic input validation.