User can now create new books and also view existing books
When user logs in, all the user related book data is downloaded from Firebase and saved in Room DB. There after, user can create and add new books in the Room DB. Also, as new books are added, they are dynamically displayed in the UI. As of now, these new books are not yet updated in the Firebase servers. This feature is yet to be implemented. This sync happens once in a day or twice.