User can now create and add books directly to Firebase.
Previously, on master branch version, the single truth of data was Room DB. However, on facing challenges in that process, in the current branch the single source of truth has been shifted to Firebase itself. Maintaining the integrity, the data is updated everywhere in Firebase. When new book is created, users, sec owners are also updated with this data. However, the processing seems to be quite slow as of now. Need to make some changes in it.