User can now see all the books.
User can now see all the owned books created by them and also all the books that are shared with them. Owned books are shown without any SHARE ICON. And the shared books are shown a SHARE ICON to differentiate between the two types. There is a bug however. User is able to add themseleves as the secondary owner. Because of this, unexpected UI is shown when displaying books. For example, their own books are shown as shared books. This is a major confusion. The bug will be resolved in the upcoming commits.