Completed book update module.
Users can now update their book data such as name and the secondary owners and their access. The updated data is synchronised with Firebase servers as well. For quick access of data, it is even saved in local caches such as Shared preferences and SQLite DB. There are many un-optimised ways of getting and updating the data. These will be optimised in the end.