Refactored code for Add book module.
The code now runs 3X faster than the previous version. This is because the current algorithm makes most of the on-device local storage facilities. Also internet data consumption reduced by 5X with the new algorithm. Also the display of books is faster than before because of the same reasons. Similar refactoring will be done for other modules as well.