Authentication module refactored.
Now users can even enter manual OTP in order to login. This feature was not available previously in this project. Although the code performs with same efficiency, there are a lot of internal code refactorings done in this module. For example, the entire code including the firebase code was put in single class previously. Now, it has been refactored and follows the Android's recommended app lifecycle pattern.