Profile picture feature finished.
The profile picture is now uploaded and downloaded from Firebase Storage. The profile picture is saved on device memory for easier access when there is no internet connection. This data is restricted only for this user. Once the user has logged out, that user's private data including profile pic would be erased.