Bog card display module completed.
Now the user can add multiple cards and view all of them by swiping across the screen. Custom imageviews are used for more user experience. They give more real time photo frame look and feel. On the back side of the card, there are other details of the card such as location, date, description. More UI polishing need to be done. But the functionality works well with only a few bugs such as over sized images. This will be resolved in the future.