Card update module 80% completed.
User can now select a card and edit it. They can add or remove photos, change any other data such as description and location of the card. The details are reflected both in server and local DB. However there are UI bug fixed that need to be done yet. They will be fixed by the next commit.