Card sync from firebase to device completed.
When a user logs in, the card data along with books and profile was not synced with firebase till now. The card data was persisted only locally in internal storage of the device. The code for syncing is now developed and works fine. There are bug fixes that need to be done.