Coding standards implemented for card module.
Hardcoded values are removed and replaced with variables. Also, resolved a bug in card display module. Previously, while the images are being downloaded from firebase, they are not dynamically updated in the UI. User had to go back and come again to see the updated UI. But now, this bug has been resolved. UI is dynamically udpated as the data is downloaded from firebase.