Modified the user profile module completely.
Now, the UI is simpler and more user friendly. User can edit their details and view them in just one screen unlike the previous UI. User does not even need to press any save button explicitly. As the details are changed, they are updated in Firebase and Local cache. The design is inspired from that of WhatsApp.