Settings module completed.
User can now change the theme in the settings screen. Also the logout option has been moved into this screen. Changes have been made to the UI by removing the navigation drawer and replacing it with tab layout at the bottom. The reason behind this change is the lack of enough items to put in the navigation drawer.