Fixed 3 bugs in the UI experience.
When a book is edited and updated, the gallery of small cards was not visible due to new array creation. This was fixed. Secondly, edit and remove options were visible for cards that are not owned, but shared with the user. This is a major security issue and has been resolved. Another minor UI bug thouogh not critical was resolved.